change timetag impl

+++ include/kernel/defs.h
@@ -145,7 +145,7 @@ int             fetchaddr(uint64, uint64*);
 void            syscall();
 
 // trap.c
-extern uint     ticks;
+extern uint64   ticks;
 void            trapinit(void);
 void            trapinithart(void);
 extern struct spinlock tickslock;
